Lab-1- Explore Closed Source Models/requirements.txt
openai==2.8.1
anthropic==0.75.0
google-genai==1.52.0
python-dotenv==1.2.1

